(PUP-2290) Issue warning when rest/find results in 404
This makes the indirector for rest.find issue a warning when
the server returns a 404. The warning is issued once.

The intent of this is to give the user more feedback about the cause of
a "not found", typically when using a non existing environment. If no
warning is given no clue is presented except an info level log item on
the master side.