(maint) Adjust overlapping test case to use environment.conf
And a non-interpolated basemodulepath.  An interpolated basemodulepath
does not work with directory environments, because it is added to the
loader earlier on, and interpolated with whatever the default
environment is.  A later call with a different environment will be stuck
with the default basemodulepath interpolation.  Not being able to
interpolate basemodulepath in a useful manor is not a problem because
you simply set the expected path in your directory environment's
environment.conf modulepath.