(PUP-2395) Update puppetmaster-passenger for httpd 2.4
In httpd 2.4 some of the conventions around a2ensite changed which broke
the postinst and postrm scripts. This commit brings in several changes from the
debian puppet packaging that have addressed this problem. An
apache2_puppetmaster_sitename function is used to determine the correct
sitename to use in sites-available. Thanks to Stig Sandbeck Mathisen for
the work in the debian puppet packaging.