(PUP-2484) Revert to merging Modulefile and metadata.json in PMT build.
Previously, if you tried to build a module that contained both a
Modulefile and a metadata.json file, the behavior of the build action
was to ignore the Modulefile and simply use the metadata.json as is.

This represented a change in behavior from earlier, where we would merge
the contents of those two files. While this is in keeping with the
direction the tool is moving, there is a small subset of users who would
be negatively impacted by the sudden change in behavior. For this reason
we are reverting to the original behavior and passing it through a
proper deprecation cycle.