(maint) Include 'f' prefix for fedora paths in the builds repo
Recent changes to acceptance install_utils fetch all repository data
to the coordinator node and then push that data to the test nodes rather
than expecting the test nodes to be able to reach our internal builds
server.  Two paths being used to fetch repo data were missing the
necessiry of including a 'f' in the version segment of the path when
fetching fedora repo data. This patch fixes this.