(PUP-2369) Ignore the 'of type Class' output added to property changes
Acceptance tests had two cases which were checking the output of
property changes, and these were tripping over the new additoin of 'of
type <class>' added in PUP-2369.  This patch just changes the regex to
ignore that section of the output so we are just checking the values,
since the class info is more for developer debugging.