(PUP-1028) Improve grammar for node expression to accept numbers.
It was previously not possible to enter a name consisting of just
a number, or where one name part was a number.

e.g.
1.2.3.4
hello.1

This grammar change fixes this problem. Note that any segment must be a
valid name or valid numeric string. Also note that white space is
ignored.