(PUP-1028) Remove pending tests that aren't needed
The new parser has been updated to be able to handle nodes with numeric
names and so the test for that case now passes for both the current and
future parser. The only remaining case is when the name is an invalid
number. The future parser won't ever parse this correctly and so the
test has been changed to assert the desired behavior, in which the
parser fails.

The new changes to the parser also allows it to handle dotted names and
so a new test has been added for that.