(PUP-2519) Create fallback environment only for default production
The goal here was to only produce a fallback environment for the default
'production' environment which would only be found if the production
directory environment did not exist.  All other cases would raise an
error during application start up if the environment did not exist.
Except for applications in agent mode which currently need the concept of
the environment they will be requesting, but should not have to create
or verify the existence of a locally matching environment in their
environmentpath.

A few other changes are made to further narrow the scope of code making
open requests for the configured environment:

* indirection request - looks up current environment now, relying on
  this having been set by whatever application is making the request.
* puppet apply - instead of creating a new loader with a single static
  environment, it overrides current_environment instead.
* puppet util commandline - does not attempt to look for additional
  plugins if there is no environment to search in.  It relies on
  application to raise an error instead, if necessary.