(PUP-2802) Add slot support for portage package provider
Portage allows installing multiple versions of a package using slots.
To support slotted packages, Puppet needs to check only the packages
installed in the specified slot, as well as install using the slot
specifier. Other package managers have the slot as part of the package
name, but portage separates this its package name syntax.
This commit adds support for checking and installing packages with slot
specifiers, allowing one to keep multiple versions of a package
up-to-date.