(PUP-2577) Use versioncmp to test Mac OS X compatibility
Prior to this commit, the version comparison which indicates
Puppet's support for various versions of Mac OS X used a
string comparison against 10.5 (the oldest version we technically
run on).  This caused versions such as "10.10" to be considered
lower than "10.5", which is incorrect.

This commit adjusts this comparison to use the same versioncmp
helper method that other methods in directoryservices.rb use to
ascertain the running version of OS X.