(PUP-2159) Serialize node.environment as environment_name
Puppet::Node used to implicitly lookup and set an environment instance
when deserializing it's environment attribute String.  On an agent run,
with directory environments set, this was likely to produce a node with
a nil environment.

Henrik Lindberg and I changed node to deserialize the environment to an
environment_name attribute, separating the process of deserialization
from lookup and assignment of an actual environment.  The configurer can
then check to see whether the node already has an environment, and if
not, if it has an environment_name, it can create a remote environment
instance for the node so the agent can proceed without needing a
directory environment to exist.