(PUP-2519) Set environment for request or fail explicitly
The Indirector::Request#environment= setter would fall back to looking
up an environment from the environment loaders if given something that
was not an actual environment instance. If directory environments have
been set, this can return nil, and while a Request#environment will fall
back to the current_environment, there was no warning that you have
attempted to set an environment and been ignored.

To make this clearer, the setter will now first attempt to use the
current_environment if it matches the given environment name, and will
explicitly raise an EnvironmentNotFound error if no matching environment
is found. The use of the current_environment with a matching name is
preferred over looking up a configured environment because there may be
no directory environment to load.